Lazy load Calendar component
- Load from separate chunk/bundle
