+++ src/utils/LazyLoader/Calendar.js
@@ -1,0 +1,1 @@
+export default () => import(/* webpackChunkName: "calendar" */ '../../fragments/Calendar');
